requirements.txt
appdirs
boto
boto3
botocore
bz2file
certifi
chardet
cycler
Django
django-extensions
docutils
gensim
idna
jmespath
kiwisolver
matplotlib
numpy
pandas
packaging
pyparsing
python-dateutil
pytz
requests
s3transfer
scipy
seaborn
six
sklearn
smart-open
stop-words
tqdm
urllib3
pyLDAvis
psycopg2-binary
gdown>=4.7.1

